+++ admin/exchange_message_traces_item_recipients_count_request_builder.go
@@ -44,9 +44,9 @@ func NewExchangeMessageTracesItemRecipientsCountRequestBuilder(rawUrl string, re
 }
 // Get get the number of the resource
 // Deprecated: Private preview for Import Export APIs as of 2021-08/PrivatePreview:importExport on 2021-08-19 and will be removed 2021-11-15
-// returns a *int64 when successful
+// returns a *int32 when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-func (m *ExchangeMessageTracesItemRecipientsCountRequestBuilder) Get(ctx context.Context, requestConfiguration *ExchangeMessageTracesItemRecipientsCountRequestBuilderGetRequestConfiguration)(*int64, error) {
+func (m *ExchangeMessageTracesItemRecipientsCountRequestBuilder) Get(ctx context.Context, requestConfiguration *ExchangeMessageTracesItemRecipientsCountRequestBuilderGetRequestConfiguration)(*int32, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
         return nil, err
@@ -54,14 +54,14 @@ func (m *ExchangeMessageTracesItemRecipientsCountRequestBuilder) Get(ctx context
     errorMapping := i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.ErrorMappings {
         "XXX": i20a3050780ee0b0cde0a884a4f35429a20d60067e3bcda382ec5400079147459.CreateODataErrorFromDiscriminatorValue,
     }
-    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int64", errorMapping)
+    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int32", errorMapping)
     if err != nil {
         return nil, err
     }
     if res == nil {
         return nil, nil
     }
-    return res.(*int64), nil
+    return res.(*int32), nil
 }
 // ToGetRequestInformation get the number of the resource
 // Deprecated: Private preview for Import Export APIs as of 2021-08/PrivatePreview:importExport on 2021-08-19 and will be removed 2021-11-15
