+++ devicemanagement/virtual_endpoint_snapshots_get_storage_accounts_with_subscription_id_request_builder.go
@@ -22,9 +22,9 @@ type VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilderG
     // Search items by search phrases
     Search *string `uriparametername:"%24search"`
     // Skip the first n items
-    Skip *int64 `uriparametername:"%24skip"`
+    Skip *int32 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilderGetRequestConfiguration struct {
@@ -55,6 +55,9 @@ func NewVirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuild
 // Deprecated: This method is obsolete. Use GetAsGetStorageAccountsWithSubscriptionIdGetResponse instead.
 // returns a VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/cloudpcsnapshot-getstorageaccounts?view=graph-rest-beta
 func (m *VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilder) Get(ctx context.Context, requestConfiguration *VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilderGetRequestConfiguration)(VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
@@ -75,6 +78,9 @@ func (m *VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuil
 // GetAsGetStorageAccountsWithSubscriptionIdGetResponse list all storage accounts cloudPcForensicStorageAccount that can be used to store a snapshot or snapshots of a Cloud PC for forensic analysis.
 // returns a VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdGetResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/cloudpcsnapshot-getstorageaccounts?view=graph-rest-beta
 func (m *VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilder) GetAsGetStorageAccountsWithSubscriptionIdGetResponse(ctx context.Context, requestConfiguration *VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdRequestBuilderGetRequestConfiguration)(VirtualEndpointSnapshotsGetStorageAccountsWithSubscriptionIdGetResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
